more compile warnings fixed for 64 bits
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@3333 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/cfgutils/cfgutils.c
@@ -210,7 +210,7 @@ error:
 
 static int set_prob(struct sip_msg *bar, char *percent_par, char *foo) 
 {
-	*probability=(int) percent_par;
+	*probability=(int)(long)percent_par;
 	return 1;
 }
